Delay CLI handler module loading
Static command descriptors now import parser-only modules where available
and load heavier command handlers only when the selected command runs.
This keeps parse-only runner imports compatible with Node 20, where
node:sqlite is not available.

https://github.com/fedify-dev/fedify/pull/808#discussion_r3422845846
